# ci: add benchmark-regression-check