--- a/.travis.yml
+++ b/.travis.yml
@@ -13,3 +13,4 @@ go:
 
 script:
   - make test
+  - bash ./.travis/benchmark-regressions-check.sh