@nodece 2020-07-22
The computes and compares statistics about benchmarks for 96dfdc9f16e41c8f9eb4f1f5ea61be1b3009190d by [benchstat](https://github.com/golang/perf/tree/master/cmd/benchstat) with golang 1.14:

```
name                                old time/op  new time/op  delta
CachedRaw-2                         30.9ns ± 1%  30.7ns ± 0%  -0.58%  (p=0.032 n=5+5)
CachedBasicModel-2                   260ns ± 3%   261ns ± 2%    ~     (p=0.508 n=5+5)
CachedRBACModel-2                    263ns ± 1%   263ns ± 2%    ~     (p=0.683 n=5+5)
CachedRBACModelSmall-2               263ns ± 2%   262ns ± 2%    ~     (p=0.873 n=5+5)
CachedRBACModelMedium-2              286ns ± 3%   291ns ± 3%    ~     (p=0.198 n=5+5)
CachedRBACModelLarge-2               274ns ± 3%   279ns ± 1%    ~     (p=0.119 n=5+5)
CachedRBACModelWithResourceRoles-2   259ns ± 1%   260ns ± 2%    ~     (p=0.611 n=5+5)
CachedRBACModelWithDomains-2         274ns ± 2%   275ns ± 2%    ~     (p=0.587 n=5+5)
CachedABACModel-2                   7.74µs ± 0%  7.67µs ± 1%    ~     (p=0.056 n=5+5)
CachedKeyMatchModel-2                288ns ± 1%   281ns ± 0%  -2.45%  (p=0.016 n=5+4)
CachedRBACModelWithDeny-2            262ns ± 3%   258ns ± 0%    ~     (p=0.103 n=5+5)
CachedPriorityModel-2                264ns ± 1%   266ns ± 5%    ~     (p=0.722 n=5+5)
CachedRBACModelMediumParallel-2      250ns ± 2%   243ns ± 2%    ~     (p=0.063 n=5+5)
HasPolicySmall-2                     975ns ± 4%   965ns ± 2%    ~     (p=0.548 n=5+5)
HasPolicyMedium-2                   1.01µs ± 3%  0.96µs ± 1%  -5.46%  (p=0.008 n=5+5)
HasPolicyLarge-2                    1.02µs ± 1%  0.97µs ± 1%  -4.65%  (p=0.008 n=5+5)
AddPolicySmall-2                    1.03µs ± 1%  0.98µs ± 1%  -5.23%  (p=0.008 n=5+5)
AddPolicyMedium-2                   1.33µs ± 2%  1.22µs ± 7%  -8.09%  (p=0.008 n=5+5)
AddPolicyLarge-2                    1.94µs ± 8%  1.89µs ± 4%    ~     (p=0.421 n=5+5)
RemovePolicySmall-2                  995ns ± 4%   965ns ± 3%    ~     (p=0.087 n=5+5)
RemovePolicyMedium-2                1.10µs ± 3%  1.09µs ± 2%    ~     (p=0.460 n=5+5)
RemovePolicyLarge-2                 2.21µs ± 4%  2.24µs ± 5%    ~     (p=0.548 n=5+5)
Raw-2                               30.5ns ± 0%  30.8ns ± 2%    ~     (p=0.095 n=5+5)
BasicModel-2                        16.9µs ±10%  16.3µs ± 1%    ~     (p=0.310 n=5+5)
RBACModel-2                         23.2µs ± 1%  23.2µs ± 2%    ~     (p=0.548 n=5+5)
RBACModelSmall-2                     203µs ± 2%   202µs ± 2%    ~     (p=1.000 n=5+5)
RBACModelMedium-2                   2.04ms ± 3%  2.03ms ± 1%    ~     (p=0.548 n=5+5)
RBACModelLarge-2                    23.9ms ± 3%  23.6ms ± 3%    ~     (p=0.548 n=5+5)
RBACModelWithResourceRoles-2        26.4µs ± 1%  26.6µs ± 1%    ~     (p=0.056 n=5+5)
RBACModelWithDomains-2              32.7µs ± 2%  32.8µs ± 1%    ~     (p=1.000 n=5+5)
ABACModel-2                         7.62µs ± 1%  7.77µs ± 1%  +2.03%  (p=0.008 n=5+5)
KeyMatchModel-2                     29.0µs ± 1%  28.8µs ± 1%    ~     (p=0.095 n=5+5)
RBACModelWithDeny-2                 24.5µs ± 2%  24.4µs ± 2%    ~     (p=1.000 n=5+5)
PriorityModel-2                     18.0µs ± 1%  18.0µs ± 1%    ~     (p=0.516 n=5+5)
RoleManagerSmall-2                  71.6µs ± 1%  71.7µs ± 1%    ~     (p=1.000 n=5+5)
RoleManagerMedium-2                  773µs ± 2%   777µs ± 1%    ~     (p=0.690 n=5+5)
RoleManagerLarge-2                  9.93ms ± 5%  9.87ms ± 5%    ~     (p=0.548 n=5+5)
```