@nodece 2020-07-22
The computes and compares statistics about benchmarks for f0457b3e78ddb73801ac16071073064688976dde by [benchstat](https://github.com/golang/perf/tree/master/cmd/benchstat) with golang 1.14:

```
name                                old time/op  new time/op  delta
CachedRaw-2                         22.5ns ± 1%  22.3ns ± 1%    ~     (p=0.079 n=5+5)
CachedBasicModel-2                   270ns ± 5%   257ns ± 1%  -4.67%  (p=0.008 n=5+5)
CachedRBACModel-2                    266ns ± 1%   258ns ± 1%  -3.09%  (p=0.008 n=5+5)
CachedRBACModelSmall-2               267ns ± 1%   261ns ± 1%  -2.10%  (p=0.016 n=5+5)
CachedRBACModelMedium-2              307ns ± 2%   290ns ± 0%  -5.41%  (p=0.008 n=5+5)
CachedRBACModelLarge-2               299ns ± 2%   284ns ± 2%  -5.08%  (p=0.008 n=5+5)
CachedRBACModelWithResourceRoles-2   267ns ± 1%   258ns ± 0%  -3.15%  (p=0.008 n=5+5)
CachedRBACModelWithDomains-2         284ns ± 2%   278ns ± 1%  -2.32%  (p=0.016 n=5+5)
CachedABACModel-2                   7.37µs ± 2%  7.10µs ± 1%  -3.67%  (p=0.008 n=5+5)
CachedKeyMatchModel-2                305ns ± 1%   308ns ± 1%  +0.92%  (p=0.024 n=5+5)
CachedRBACModelWithDeny-2            261ns ± 1%   257ns ± 0%  -1.61%  (p=0.000 n=5+4)
CachedPriorityModel-2                257ns ± 1%   258ns ± 0%    ~     (p=0.103 n=5+5)
CachedRBACModelMediumParallel-2      320ns ± 2%   320ns ± 6%    ~     (p=0.587 n=5+5)
HasPolicySmall-2                     875ns ± 1%   860ns ± 2%    ~     (p=0.056 n=5+5)
HasPolicyMedium-2                    867ns ± 2%   860ns ± 1%    ~     (p=0.548 n=5+5)
HasPolicyLarge-2                     882ns ± 2%   874ns ± 1%    ~     (p=0.278 n=5+5)
AddPolicySmall-2                     875ns ± 1%   868ns ± 1%    ~     (p=0.056 n=5+5)
AddPolicyMedium-2                   1.22µs ± 3%  1.20µs ± 2%    ~     (p=0.548 n=5+5)
AddPolicyLarge-2                    1.87µs ± 5%  1.94µs ±10%    ~     (p=0.151 n=5+5)
RemovePolicySmall-2                  886ns ± 1%   880ns ± 1%    ~     (p=0.595 n=5+5)
RemovePolicyMedium-2                 960ns ± 2%   951ns ± 2%    ~     (p=0.198 n=5+5)
RemovePolicyLarge-2                 1.92µs ± 6%  1.94µs ± 3%    ~     (p=0.548 n=5+5)
Raw-2                               22.1ns ± 0%  22.1ns ± 1%    ~     (p=1.000 n=4+5)
BasicModel-2                        14.9µs ± 0%  15.0µs ± 0%    ~     (p=0.548 n=5+5)
RBACModel-2                         21.3µs ± 0%  21.1µs ± 0%  -1.12%  (p=0.008 n=5+5)
RBACModelSmall-2                     182µs ± 1%   180µs ± 1%  -1.36%  (p=0.032 n=5+5)
RBACModelMedium-2                   1.98ms ± 1%  1.95ms ± 2%    ~     (p=0.095 n=5+5)
RBACModelLarge-2                    25.9ms ± 3%  24.7ms ± 3%  -4.64%  (p=0.016 n=5+5)
RBACModelWithResourceRoles-2        25.2µs ± 1%  24.8µs ± 0%  -1.65%  (p=0.008 n=5+5)
RBACModelWithDomains-2              30.3µs ± 1%  30.0µs ± 1%  -0.91%  (p=0.016 n=5+5)
ABACModel-2                         7.13µs ± 1%  7.08µs ± 0%  -0.79%  (p=0.016 n=5+5)
KeyMatchModel-2                     27.4µs ± 1%  27.3µs ± 1%    ~     (p=0.151 n=5+5)
RBACModelWithDeny-2                 22.3µs ± 0%  22.2µs ± 1%    ~     (p=0.151 n=5+5)
PriorityModel-2                     16.7µs ± 1%  16.6µs ± 0%  -0.69%  (p=0.008 n=5+5)
RoleManagerSmall-2                  62.8µs ± 1%  61.9µs ± 0%  -1.49%  (p=0.008 n=5+5)
RoleManagerMedium-2                  679µs ± 1%   669µs ± 1%  -1.56%  (p=0.008 n=5+5)
RoleManagerLarge-2                  11.2ms ± 3%  10.8ms ± 1%  -3.28%  (p=0.032 n=5+5)
```