@nodece 2020-07-22
The computes and compares statistics about benchmarks for 88b19f8cc8fe026e16cbe0feb4cab402e99cc4f7 by [benchstat](https://github.com/golang/perf/tree/master/cmd/benchstat) with golang 1.14:

```
name                                old time/op  new time/op  delta
CachedRaw-2                         32.4ns ± 7%  32.0ns ± 1%    ~     (p=0.905 n=5+4)
CachedBasicModel-2                   270ns ± 3%   271ns ± 3%    ~     (p=0.730 n=5+5)
CachedRBACModel-2                    267ns ± 1%   269ns ± 4%    ~     (p=0.984 n=5+5)
CachedRBACModelSmall-2               281ns ± 5%   284ns ± 4%    ~     (p=1.000 n=5+5)
CachedRBACModelMedium-2              296ns ± 4%   294ns ± 2%    ~     (p=0.984 n=5+5)
CachedRBACModelLarge-2               303ns ± 2%   306ns ± 7%    ~     (p=0.802 n=5+5)
CachedRBACModelWithResourceRoles-2   277ns ± 9%   282ns ± 6%    ~     (p=0.381 n=5+5)
CachedRBACModelWithDomains-2         283ns ± 5%   288ns ± 2%    ~     (p=0.151 n=5+5)
CachedABACModel-2                   8.13µs ± 3%  8.33µs ±12%    ~     (p=1.000 n=5+5)
CachedKeyMatchModel-2                303ns ±11%   299ns ± 4%    ~     (p=0.548 n=5+5)
CachedRBACModelWithDeny-2            270ns ± 3%   274ns ± 3%    ~     (p=0.341 n=5+5)
CachedPriorityModel-2                263ns ± 2%   266ns ± 2%    ~     (p=0.119 n=5+5)
CachedRBACModelMediumParallel-2      261ns ± 4%   264ns ± 2%    ~     (p=0.389 n=5+5)
HasPolicySmall-2                    1.06µs ± 6%  1.02µs ± 4%    ~     (p=0.341 n=5+5)
HasPolicyMedium-2                   1.04µs ± 5%  1.02µs ± 1%    ~     (p=0.421 n=5+5)
HasPolicyLarge-2                    1.04µs ± 2%  1.10µs ± 7%  +5.78%  (p=0.008 n=5+5)
AddPolicySmall-2                    1.06µs ± 0%  1.07µs ± 3%    ~     (p=0.254 n=4+5)
AddPolicyMedium-2                   1.30µs ± 4%  1.34µs ± 1%    ~     (p=0.135 n=5+5)
AddPolicyLarge-2                    2.03µs ± 5%  2.09µs ± 8%    ~     (p=0.548 n=5+5)
RemovePolicySmall-2                 1.06µs ± 7%  1.04µs ± 1%    ~     (p=0.548 n=5+5)
RemovePolicyMedium-2                1.14µs ± 4%  1.18µs ±14%    ~     (p=0.841 n=5+5)
RemovePolicyLarge-2                 2.37µs ±10%  2.40µs ± 7%    ~     (p=1.000 n=5+5)
Raw-2                               32.7ns ± 6%  32.0ns ± 6%    ~     (p=0.548 n=5+5)
BasicModel-2                        17.2µs ± 4%  17.6µs ± 8%    ~     (p=0.421 n=5+5)
RBACModel-2                         24.6µs ± 3%  25.3µs ± 5%    ~     (p=0.310 n=5+5)
RBACModelSmall-2                     224µs ± 4%   219µs ± 2%    ~     (p=0.095 n=5+5)
RBACModelMedium-2                   2.18ms ± 2%  2.21ms ± 3%    ~     (p=0.548 n=5+5)
RBACModelLarge-2                    27.1ms ± 5%  28.6ms ± 8%    ~     (p=0.222 n=5+5)
RBACModelWithResourceRoles-2        27.8µs ± 2%  29.2µs ± 3%  +5.10%  (p=0.008 n=5+5)
RBACModelWithDomains-2              35.7µs ± 9%  35.0µs ± 2%    ~     (p=1.000 n=5+5)
ABACModel-2                         8.06µs ± 2%  8.65µs ± 6%  +7.34%  (p=0.008 n=5+5)
KeyMatchModel-2                     31.1µs ±15%  31.0µs ± 2%    ~     (p=0.151 n=5+5)
RBACModelWithDeny-2                 26.1µs ± 7%  26.3µs ± 3%    ~     (p=0.548 n=5+5)
PriorityModel-2                     18.5µs ± 1%  19.5µs ±11%  +5.33%  (p=0.032 n=5+5)
RoleManagerSmall-2                  77.2µs ± 6%  79.5µs ± 5%    ~     (p=0.310 n=5+5)
RoleManagerMedium-2                  845µs ± 4%   901µs ±11%    ~     (p=0.056 n=5+5)
RoleManagerLarge-2                  10.6ms ± 2%  11.0ms ±10%    ~     (p=0.690 n=5+5)
```