@nodece 2020-07-22
The computes and compares statistics about benchmarks for 2e388cbbe0bfcf531b49f184c928d4ddbcef76c8 by [benchstat](https://github.com/golang/perf/tree/master/cmd/benchstat) with golang 1.14:

```
name                                old time/op  new time/op  delta
CachedRaw-2                         23.2ns ± 4%  23.6ns ± 1%    ~     (p=0.135 n=5+5)
CachedBasicModel-2                   273ns ± 1%   279ns ± 1%  +2.12%  (p=0.016 n=5+5)
CachedRBACModel-2                    275ns ± 1%   275ns ± 1%    ~     (p=0.984 n=5+5)
CachedRBACModelSmall-2               280ns ± 1%   275ns ± 1%  -1.86%  (p=0.024 n=5+5)
CachedRBACModelMedium-2              318ns ± 2%   302ns ± 1%  -5.03%  (p=0.008 n=5+5)
CachedRBACModelLarge-2               317ns ± 4%   308ns ± 2%    ~     (p=0.079 n=5+5)
CachedRBACModelWithResourceRoles-2   272ns ± 1%   273ns ± 2%    ~     (p=0.508 n=5+5)
CachedRBACModelWithDomains-2         292ns ± 1%   290ns ± 1%    ~     (p=0.270 n=5+5)
CachedABACModel-2                   7.53µs ± 1%  7.50µs ± 2%    ~     (p=0.421 n=5+5)
CachedKeyMatchModel-2                321ns ± 2%   323ns ± 2%    ~     (p=0.754 n=5+5)
CachedRBACModelWithDeny-2            269ns ± 1%   273ns ± 2%    ~     (p=0.206 n=5+5)
CachedPriorityModel-2                269ns ± 1%   280ns ± 1%  +4.17%  (p=0.008 n=5+5)
CachedRBACModelMediumParallel-2      343ns ± 4%   352ns ± 4%    ~     (p=0.087 n=5+5)
HasPolicySmall-2                     918ns ± 2%   929ns ± 2%    ~     (p=0.206 n=5+5)
HasPolicyMedium-2                    921ns ± 1%   925ns ± 1%    ~     (p=0.587 n=5+5)
HasPolicyLarge-2                     949ns ± 2%   967ns ± 1%  +1.92%  (p=0.032 n=5+5)
AddPolicySmall-2                     928ns ± 1%   953ns ± 2%  +2.70%  (p=0.016 n=5+5)
AddPolicyMedium-2                   1.34µs ± 2%  1.35µs ± 1%    ~     (p=1.000 n=5+5)
AddPolicyLarge-2                    2.07µs ± 6%  2.08µs ± 6%    ~     (p=1.000 n=5+5)
RemovePolicySmall-2                  940ns ± 2%   964ns ± 2%  +2.53%  (p=0.032 n=5+5)
RemovePolicyMedium-2                1.02µs ± 4%  1.05µs ± 2%    ~     (p=0.310 n=5+5)
RemovePolicyLarge-2                 2.06µs ± 2%  2.19µs ± 4%  +6.00%  (p=0.008 n=5+5)
Raw-2                               23.0ns ± 1%  23.4ns ± 1%  +1.82%  (p=0.016 n=5+5)
BasicModel-2                        15.9µs ± 1%  16.3µs ± 1%  +2.79%  (p=0.008 n=5+5)
RBACModel-2                         22.4µs ± 1%  23.3µs ± 1%  +4.10%  (p=0.008 n=5+5)
RBACModelSmall-2                     193µs ± 2%   197µs ± 3%    ~     (p=0.095 n=5+5)
RBACModelMedium-2                   2.15ms ± 1%  2.25ms ± 2%  +4.57%  (p=0.008 n=5+5)
RBACModelLarge-2                    27.9ms ± 3%  29.1ms ± 2%  +4.01%  (p=0.032 n=5+5)
RBACModelWithResourceRoles-2        27.0µs ± 1%  28.6µs ± 3%  +5.88%  (p=0.008 n=5+5)
RBACModelWithDomains-2              32.7µs ± 2%  32.0µs ± 0%    ~     (p=0.175 n=5+4)
ABACModel-2                         7.70µs ± 1%  7.52µs ± 1%  -2.28%  (p=0.008 n=5+5)
KeyMatchModel-2                     30.2µs ± 3%  28.9µs ± 2%  -4.39%  (p=0.008 n=5+5)
RBACModelWithDeny-2                 25.1µs ± 3%  23.9µs ± 3%  -4.53%  (p=0.016 n=5+5)
PriorityModel-2                     17.8µs ± 1%  17.6µs ± 1%    ~     (p=0.063 n=5+5)
RoleManagerSmall-2                  65.6µs ± 1%  65.9µs ± 3%    ~     (p=0.841 n=5+5)
RoleManagerMedium-2                  722µs ± 2%   731µs ± 1%    ~     (p=0.056 n=5+5)
RoleManagerLarge-2                  12.2ms ± 3%  12.3ms ± 4%    ~     (p=0.690 n=5+5)
```