@nodece 2020-07-22
The computes and compares statistics about benchmarks for 17d18263cdfa73d303e1e9a4f908934687f5f3e0 by [benchstat](https://github.com/golang/perf/tree/master/cmd/benchstat) with golang 1.14:

```
name                                old time/op  new time/op  delta
CachedRaw-2                         27.0ns ± 3%  26.6ns ± 2%     ~     (p=0.167 n=5+5)
CachedBasicModel-2                   245ns ± 2%   244ns ± 2%     ~     (p=0.952 n=5+5)
CachedRBACModel-2                    244ns ± 3%   248ns ± 1%     ~     (p=0.143 n=5+5)
CachedRBACModelSmall-2               243ns ± 2%   251ns ± 3%   +3.37%  (p=0.008 n=5+5)
CachedRBACModelMedium-2              261ns ± 3%   260ns ± 3%     ~     (p=0.802 n=5+5)
CachedRBACModelLarge-2               251ns ± 2%   258ns ± 3%     ~     (p=0.095 n=5+5)
CachedRBACModelWithResourceRoles-2   245ns ± 2%   251ns ± 1%   +2.28%  (p=0.024 n=5+5)
CachedRBACModelWithDomains-2         256ns ± 1%   263ns ± 2%   +2.74%  (p=0.016 n=5+5)
CachedABACModel-2                   7.00µs ± 1%  7.17µs ± 1%   +2.44%  (p=0.008 n=5+5)
CachedKeyMatchModel-2                272ns ± 2%   280ns ± 2%   +2.79%  (p=0.040 n=5+5)
CachedRBACModelWithDeny-2            240ns ± 1%   251ns ± 2%   +4.84%  (p=0.008 n=5+5)
CachedPriorityModel-2                240ns ± 1%   253ns ± 1%   +5.16%  (p=0.008 n=5+5)
CachedRBACModelMediumParallel-2      243ns ± 3%   272ns ±10%  +12.11%  (p=0.008 n=5+5)
HasPolicySmall-2                     897ns ± 2%  1057ns ± 3%  +17.91%  (p=0.008 n=5+5)
HasPolicyMedium-2                    930ns ± 2%  1086ns ± 4%  +16.82%  (p=0.008 n=5+5)
HasPolicyLarge-2                     928ns ± 2%  1112ns ± 4%  +19.80%  (p=0.008 n=5+5)
AddPolicySmall-2                     926ns ± 3%  1082ns ± 5%  +16.82%  (p=0.008 n=5+5)
AddPolicyMedium-2                   1.12µs ± 2%  1.34µs ± 1%  +19.61%  (p=0.008 n=5+5)
AddPolicyLarge-2                    1.83µs ± 6%  2.09µs ± 9%  +14.40%  (p=0.016 n=5+5)
RemovePolicySmall-2                  922ns ± 3%  1107ns ± 1%  +20.04%  (p=0.008 n=5+5)
RemovePolicyMedium-2                1.01µs ± 4%  1.22µs ± 4%  +19.97%  (p=0.008 n=5+5)
RemovePolicyLarge-2                 2.01µs ± 4%  2.46µs ± 9%  +22.63%  (p=0.008 n=5+5)
Raw-2                               26.3ns ± 4%  30.9ns ± 3%  +17.46%  (p=0.008 n=5+5)
BasicModel-2                        14.7µs ± 1%  17.1µs ± 3%  +16.58%  (p=0.008 n=5+5)
RBACModel-2                         21.2µs ± 2%  24.7µs ± 3%  +16.48%  (p=0.008 n=5+5)
RBACModelSmall-2                     181µs ± 1%   217µs ± 3%  +19.78%  (p=0.008 n=5+5)
RBACModelMedium-2                   1.84ms ± 2%  2.23ms ± 3%  +21.20%  (p=0.008 n=5+5)
RBACModelLarge-2                    21.0ms ± 4%  24.0ms ± 7%  +13.96%  (p=0.008 n=5+5)
RBACModelWithResourceRoles-2        24.6µs ± 2%  25.1µs ± 6%     ~     (p=0.421 n=5+5)
RBACModelWithDomains-2              30.2µs ± 1%  30.2µs ± 2%     ~     (p=1.000 n=5+5)
ABACModel-2                         6.97µs ± 3%  7.28µs ± 4%   +4.53%  (p=0.008 n=5+5)
KeyMatchModel-2                     26.4µs ± 5%  26.5µs ± 2%     ~     (p=0.222 n=5+5)
RBACModelWithDeny-2                 22.3µs ± 1%  22.6µs ± 3%     ~     (p=0.310 n=5+5)
PriorityModel-2                     16.5µs ± 3%  16.4µs ± 1%     ~     (p=0.841 n=5+5)
RoleManagerSmall-2                  64.7µs ± 1%  66.2µs ± 2%   +2.30%  (p=0.008 n=5+5)
RoleManagerMedium-2                  664µs ± 2%   694µs ± 4%   +4.64%  (p=0.032 n=5+5)
RoleManagerLarge-2                  8.81ms ± 1%  8.96ms ± 4%     ~     (p=0.841 n=5+5)
```